@@ -7,8 +7,8 @@ name = "aiida_test_cache"
 
 [project]
 name = "aiida-test-cache"
+version = "0.0.1"
 description = "A pytest plugin to simplify testing of AiiDA workflows"
-dynamic = ["version"]
 authors = [
     {name = "Dominik Gresch"},
     {name = "Leopold Talirz"},
@@ -23,8 +23,6 @@ classifiers = [
     "Environment :: Plugins",
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -41,27 +39,25 @@ keywords = [
     "cache",
 ]
 urls = {Homepage = "https://aiida-testing.readthedocs.io/"}
-requires-python = ">=3.7"
-# Note the dependency on setuptools due to pkg_resources
+requires-python = ">=3.9"
 dependencies = [
     "aiida-core>=1.0.0,<3",
     "pytest>=7.0",
     "voluptuous~=0.12",
-    "setuptools",
 ]
 
 [project.optional-dependencies]
 docs = ["sphinx", "sphinx-rtd-theme"]
 tests = [
     "pgtest~=1.3.1",
-    "aiida-diff",
+    "aiida-diff~=2.0.0",
     "pytest-datadir",
     "pytest-mock",
+    "pytest-cov>=4.0",
 ]
 pre_commit = [
     "pre-commit",
     "mypy==1.13",
-    "types-setuptools==65.7.0.3",
     "types-PyYAML",
 ]
 dev = [
@@ -76,11 +72,9 @@ aiida_mock_code = "aiida_test_cache.mock_code"
 aiida_archive_cache = "aiida_test_cache.archive_cache"
 
 [tool.pytest.ini_options]
-addopts = '--durations-min=3 --strict-config --strict-markers -ra'
+addopts = '--durations=10 --durations-min=3 --strict-config --strict-markers -ra --cov-report xml --cov-append'
 filterwarnings = [
     'ignore:Creating AiiDA configuration:UserWarning:aiida',
-    # The following deprecation warnings come from Python 3.12 stdlib modules
-    "ignore:datetime.datetime.:DeprecationWarning:",
 ]
 
 [tool.yapf]
@@ -139,16 +133,16 @@ ignore_missing_imports = true
 [tool.tox]
 legacy_tox_ini = """
 [tox]
-envlist = py38
+envlist = py39
 
 [testenv]
 usedevelop = true
 
-[testenv:py{37,38,39,310,311}]
+[testenv:py{39,310,311,312}]
 extras = tests
 commands = pytest {posargs}
 
-[testenv:pre-commit-py{38,39,310,311}]
+[testenv:pre-commit-py{39,310,311,312}]
 extras = 
     pre_commit
 commands = pre-commit {posargs:run --all}
