@@ -38,7 +38,7 @@ class DiffWorkChain(WorkChain):
         return ToContext(diff_calc=running)
 
     def results(self):
-        computed_diff = self.ctx.diff_calc.get_outgoing().get_node_by_label('diff')
+        computed_diff = self.ctx.diff_calc.base.links.get_outgoing().get_node_by_label('diff')
         self.out('computed_diff', computed_diff)
 
 
@@ -61,12 +61,12 @@ def check_diff_workchain_fixture():
         assert res['computed_diff'].get_content() == EXPECTED_DIFF
 
         #Test if cache was used?
-        diffjob = node.get_outgoing().get_node_by_label('CALL')
-        cache_src = diffjob.get_cache_source()
-        print(diffjob._get_objects_to_hash())  # in case of failure to compare
+        diffjob = node.base.links.get_outgoing().get_node_by_label('CALL')
+        cache_src = diffjob.base.caching.get_cache_source()
 
-        calc_hash = diffjob.get_hash()
-        assert calc_hash == EXPECTED_HASH
+        calc_hash = diffjob.base.caching.get_hash()
+        assert calc_hash == EXPECTED_HASH, f'Hash mismatch. hashed objects: {diffjob.base.caching._get_objects_to_hash()}'
+
         #Make sure that the cache was used if it should have been
         if should_have_used_cache:
             assert cache_src is not None
@@ -135,8 +135,8 @@ def test_mock_hash_codes(aiida_profile_clean, mock_code_factory, liberal_hash):
         entry_point=CALC_ENTRY_POINT,
         ignore_paths=('_aiidasubmit.sh', 'file*')
     )
-    objs = mock_code._get_objects_to_hash()
-    assert objs == [mock_code.get_attribute(key='input_plugin')]  #, mock_code.get_computer_name()]
+    objs = mock_code.base.caching._get_objects_to_hash()
+    assert objs == [mock_code.base.attributes.get(key='input_plugin')]
 
 
 @pytest.mark.parametrize(
