@@ -37,21 +37,12 @@ def test_ignore_paths(run_directory, tmp_path_factory):
     storage_directory = tmp_path_factory.mktemp('storage')
 
     # ignore everything
-    copy_files(
-        src_dir=run_directory, dest_dir=storage_directory, ignore_files=(), ignore_paths=('*', )
-    )
+    copy_files(src_dir=run_directory, dest_dir=storage_directory, ignore_paths=('*', ))
     assert not (storage_directory / '_aiidasubmit.sh').is_file()
 
     # ignore subfolder
-    copy_files(
-        src_dir=run_directory, dest_dir=storage_directory, ignore_files=(), ignore_paths=('my/', )
-    )
-    copy_files(
-        src_dir=run_directory,
-        dest_dir=storage_directory,
-        ignore_files=(),
-        ignore_paths=('my/subfolder', )
-    )
+    copy_files(src_dir=run_directory, dest_dir=storage_directory, ignore_paths=('my/', ))
+    copy_files(src_dir=run_directory, dest_dir=storage_directory, ignore_paths=('my/subfolder', ))
     assert (storage_directory / '_aiidasubmit.sh').is_file()
     assert (storage_directory / 'file2.txt').is_file()
     assert not (storage_directory / 'my' / 'subfolder' / 'file3.txt').is_file()
@@ -60,37 +51,10 @@ def test_ignore_paths(run_directory, tmp_path_factory):
     copy_files(
         src_dir=run_directory,
         dest_dir=storage_directory,
-        ignore_files=(),
         ignore_paths=('my/subfolder/file3.txt', )
     )
     assert not (storage_directory / 'my' / 'subfolder' / 'file3.txt').is_file()
 
     # all should be there
-    copy_files(src_dir=run_directory, dest_dir=storage_directory, ignore_files=(), ignore_paths=())
-    assert (storage_directory / 'my' / 'subfolder' / 'file3.txt').is_file()
-
-
-def test_ignore_files(run_directory, tmp_path_factory):
-    """Test that ignore_files works as expected."""
-    storage_directory = tmp_path_factory.mktemp('storage')
-
-    # ignore everything
-    copy_files(
-        src_dir=run_directory, dest_dir=storage_directory, ignore_files=('*'), ignore_paths=()
-    )
-    assert not (storage_directory / '_aiidasubmit.sh').is_file()
-    assert not (storage_directory / 'my' / 'subfolder' / 'file3.txt').is_file()
-
-    # ignore only specific file (from subfolder)
-    copy_files(
-        src_dir=run_directory,
-        dest_dir=storage_directory,
-        ignore_files=('file3.txt', ),
-        ignore_paths=()
-    )
-    assert not (storage_directory / 'my' / 'subfolder' / 'file3.txt').is_file()
-    assert (storage_directory / 'file2.txt').is_file()
-
-    # all should be there
-    copy_files(src_dir=run_directory, dest_dir=storage_directory, ignore_files=(), ignore_paths=())
+    copy_files(src_dir=run_directory, dest_dir=storage_directory, ignore_paths=())
     assert (storage_directory / 'my' / 'subfolder' / 'file3.txt').is_file()
